backend/requirements.txt
fastapi==0.116.1
mangum==0.19.0
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic-settings==2.10.1
python-dotenv==1.1.1
requests==2.32.4
beautifulsoup4==4.13.4
